// src/libs/generated/clients/prefectures.ts
/**
 * Generated by Orval 🍺
 * Do not edit manually.
 * 都道府県別人口構成API
 * RESAS(地域経済分析システム)のデータを元にした、都道府県別の人口構成を取得するAPI。

 * OpenAPI spec version: 1.0.0
 * biome-ignore-all lint: due to automatically generated contents
 */
import type { Error, PrefecturesResponse } from './api.schemas';

/**
 * 全都道府県の一覧を取得します
 * @summary 都道府県一覧取得
 */
export type getPrefecturesResponse200 = {
  data: PrefecturesResponse;
  status: 200;
};

export type getPrefecturesResponse400 = {
  data: Error;
  status: 400;
};

export type getPrefecturesResponse403 = {
  data: Error;
  status: 403;
};

export type getPrefecturesResponse404 = {
  data: Error;
  status: 404;
};

export type getPrefecturesResponse500 = {
  data: Error;
  status: 500;
};

export type getPrefecturesResponseSuccess = getPrefecturesResponse200 & {
  headers: Headers;
};
export type getPrefecturesResponseError = (
  | getPrefecturesResponse400
  | getPrefecturesResponse403
  | getPrefecturesResponse404
  | getPrefecturesResponse500
) & {
  headers: Headers;
};

export type getPrefecturesResponse =
  | getPrefecturesResponseSuccess
  | getPrefecturesResponseError;

export const getGetPrefecturesUrl = () => {
  return `https://frontend-engineer-codecheck-api.mirai.yumemi.io/api/v1/prefectures`;
};

export const getPrefectures = async (
  options?: RequestInit,
): Promise<getPrefecturesResponse> => {
  const res = await fetch(getGetPrefecturesUrl(), {
    ...options,
    method: 'GET',
  });

  const body = [204, 205, 304].includes(res.status) ? null : await res.text();

  const data: getPrefecturesResponse['data'] = body ? JSON.parse(body) : {};
  return {
    data,
    status: res.status,
    headers: res.headers,
  } as getPrefecturesResponse;
};